add clean_folder_minutes

+++ pyganizer.py
@@ -5,9 +5,13 @@ import time
 parser = argparse.ArgumentParser(description="File organizer with Python.")
 subparser = parser.add_subparsers(dest="command")
 
-clean_parser = subparser.add_parser("clean_days")
-clean_parser.add_argument("folder", type=str, help="The folder destination.")
-clean_parser.add_argument("--days", type=int, default=30, help="Delete files older than this date.")
+clean_parser_days = subparser.add_parser("clean_days")
+clean_parser_days.add_argument("folder", type=str, help="The folder destination.")
+clean_parser_days.add_argument("--days", type=int, help="Delete files older than this date.")
+
+clean_parser_minutes = subparser.add_parser("clean_minutes")
+clean_parser_minutes.add_argument("folder", type=str, help="The folder destination.")
+clean_parser_minutes.add_argument("--minutes", type=int, help="Delete files older than this time in minutes.")
 
 args = parser.parse_args()
 
@@ -17,14 +21,31 @@ def clean_folder_days(folder, days):
     
     for items in os.listdir(folder):
         item_path = os.path.join(folder, items)
-        if os.path.isfile(item_path):
+        modification_date = os.path.getmtime(item_path)
+        if os.path.isfile(item_path) and modification_date <= cutoff:
             try:
                 os.remove(item_path)
                 print(f"{item_path} successfully deleted.")
             except OSError as e:
                 print(f"Trouble deleting {item_path}: {e}")        
                 
+def clean_folder_minutes(folder, minutes):
+    now = time.time()
+    cutoff = now - (minutes * 60)
+    
+    for items in os.listdir(folder):
+        item_path = os.path.join(folder, items)
+        modification_date = os.path.getmtime(item_path)
+        if os.path.isfile(item_path) and modification_date <= cutoff:
+            try:
+                os.remove(item_path)
+                print(f"{item_path} successfully deleted.")
+            except OSError as e:
+                print(f"Error deleting file {item_path}: {e}")
+                
 if args.command == "clean_days":
     clean_folder_days(args.folder, args.days)
+elif args.command == "clean_minutes":
+    clean_parser_minutes(args.folder, args.minutes)
 else:
     parser.print_help()
